+++ src/Symfony/Component/ExpressionLanguage/Tests/Node/ConstantNodeTest.php
@@ -50,6 +50,7 @@ class ConstantNodeTest extends AbstractNodeTest
             array('3', new ConstantNode(3)),
             array('3.3', new ConstantNode(3.3)),
             array('"foo"', new ConstantNode('foo')),
+            array('foo', new ConstantNode('foo', true)),
             array('{0: 1, "b": "a", 1: true}', new ConstantNode(array(1, 'b' => 'a', true))),
             array('{"a\\"b": "c", "a\\\\b": "d"}', new ConstantNode(array('a"b' => 'c', 'a\\b' => 'd'))),
             array('["c", "d"]', new ConstantNode(array('c', 'd'))),