+++ src/Symfony/Component/Translation/Catalogue/AbstractOperation.php
@@ -13,6 +13,8 @@ namespace Symfony\Component\Translation\Catalogue;
 
 use Symfony\Component\Translation\MessageCatalogue;
 use Symfony\Component\Translation\MessageCatalogueInterface;
+use Symfony\Component\Translation\Exception\InvalidArgumentException;
+use Symfony\Component\Translation\Exception\LogicException;
 
 /**
  * Base catalogues binary operation class.
@@ -72,12 +74,12 @@ abstract class AbstractOperation implements OperationInterface
      * @param MessageCatalogueInterface $source The source catalogue
      * @param MessageCatalogueInterface $target The target catalogue
      *
-     * @throws \LogicException
+     * @throws LogicException
      */
     public function __construct(MessageCatalogueInterface $source, MessageCatalogueInterface $target)
     {
         if ($source->getLocale() !== $target->getLocale()) {
-            throw new \LogicException('Operated catalogues must belong to the same locale.');
+            throw new LogicException('Operated catalogues must belong to the same locale.');
         }
 
         $this->source = $source;
@@ -105,7 +107,7 @@ abstract class AbstractOperation implements OperationInterface
     public function getMessages($domain)
     {
         if (!in_array($domain, $this->getDomains())) {
-            throw new \InvalidArgumentException(sprintf('Invalid domain: %s.', $domain));
+            throw new InvalidArgumentException(sprintf('Invalid domain: %s.', $domain));
         }
 
         if (!isset($this->messages[$domain]['all'])) {
@@ -121,7 +123,7 @@ abstract class AbstractOperation implements OperationInterface
     public function getNewMessages($domain)
     {
         if (!in_array($domain, $this->getDomains())) {
-            throw new \InvalidArgumentException(sprintf('Invalid domain: %s.', $domain));
+            throw new InvalidArgumentException(sprintf('Invalid domain: %s.', $domain));
         }
 
         if (!isset($this->messages[$domain]['new'])) {
@@ -137,7 +139,7 @@ abstract class AbstractOperation implements OperationInterface
     public function getObsoleteMessages($domain)
     {
         if (!in_array($domain, $this->getDomains())) {
-            throw new \InvalidArgumentException(sprintf('Invalid domain: %s.', $domain));
+            throw new InvalidArgumentException(sprintf('Invalid domain: %s.', $domain));
         }
 
         if (!isset($this->messages[$domain]['obsolete'])) {