+++ src/Symfony/Component/Translation/Dumper/YamlFileDumper.php
@@ -14,6 +14,7 @@ namespace Symfony\Component\Translation\Dumper;
 use Symfony\Component\Translation\MessageCatalogue;
 use Symfony\Component\Translation\Util\ArrayConverter;
 use Symfony\Component\Yaml\Yaml;
+use Symfony\Component\Translation\Exception\LogicException;
 
 /**
  * YamlFileDumper generates yaml files from a message catalogue.
@@ -28,7 +29,7 @@ class YamlFileDumper extends FileDumper
     public function formatCatalogue(MessageCatalogue $messages, $domain, array $options = array())
     {
         if (!class_exists('Symfony\Component\Yaml\Yaml')) {
-            throw new \LogicException('Dumping translations in the YAML format requires the Symfony Yaml component.');
+            throw new LogicException('Dumping translations in the YAML format requires the Symfony Yaml component.');
         }
 
         $data = $messages->all($domain);