+++ src/Symfony/Component/VarDumper/Tests/Fixtures/Twig.php
@@ -28,7 +28,12 @@ class __TwigTemplate_VarDumperFixture_u75a09 extends Twig_Template
     {
         return array (19 => 2);
     }
+
+    public function getSource()
+    {
+        return "   foo bar
+     twig source
+
+";
+    }
 }
-/*   foo bar*/
-/*     twig source*/
-/* */